Refactor step1_name
- aligned_title_text
- label_text
- signup_button
- signup_textfield
