+++ lib/widgets/signup/aligned_title_text.dart
@@ -1,0 +1,25 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_screenutil/flutter_screenutil.dart';
+
+class AlignedTitleText extends StatelessWidget {
+  final String title;
+
+  const AlignedTitleText({
+    super.key,
+    required this.title,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return Align(
+      alignment: Alignment.centerLeft,
+      child: Text(
+        title,
+        style: TextStyle(
+          fontSize: 28.sp,
+          fontWeight: FontWeight.bold,
+        ),
+      ),
+    );
+  }
+}
