+++ lib/widgets/signup/step1_name.dart
@@ -1,8 +1,11 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:provider/provider.dart';
-import 'package:workout_app/colors.dart';
 import 'package:workout_app/models/signup_model.dart';
+import 'package:workout_app/widgets/signup/aligned_title_text.dart';
+import 'package:workout_app/widgets/signup/label_text.dart';
+import 'package:workout_app/widgets/signup/signup_button.dart';
+import 'package:workout_app/widgets/signup/signup_textfield.dart';
 import 'package:workout_app/widgets/signup/step2_email.dart';
 
 class Step1Name extends StatefulWidget {
@@ -37,28 +40,6 @@ class _Step1NameState extends State<Step1Name> {
   Widget build(BuildContext context) {
     final nameController = Provider.of<SignupModel>(context).nameController;
 
-    label() {
-      if (_focusNode.hasFocus) {
-        return Text(
-          "이름",
-          style: TextStyle(
-            fontSize: 16.sp,
-            color: ColorsTheme.point,
-          ),
-        );
-      } else if (!_focusNode.hasFocus && nameController.text.isNotEmpty) {
-        return Text(
-          "이름",
-          style: TextStyle(
-            fontSize: 16.sp,
-            color: ColorsTheme.gray600,
-          ),
-        );
-      } else {
-        return null;
-      }
-    }
-
     return GestureDetector(
       onTap: () => FocusScope.of(context).unfocus(),
       child: Scaffold(
@@ -70,98 +51,47 @@ class _Step1NameState extends State<Step1Name> {
             ),
           ),
           child: Center(
-            child: Column(
-              mainAxisAlignment: MainAxisAlignment.spaceBetween,
-              children: [
-                Padding(
-                  padding: EdgeInsets.symmetric(
-                    horizontal: 25.w,
-                  ),
-                  child: Column(
+            child: Padding(
+              padding: EdgeInsets.symmetric(
+                horizontal: 25.w,
+              ),
+              child: Column(
+                mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                children: [
+                  Column(
                     children: [
                       SizedBox(
                         height: 175.h,
                       ),
-                      Align(
-                        alignment: Alignment.centerLeft,
-                        child: Text(
-                          "이름을 입력해주세요",
-                          style: TextStyle(
-                            fontSize: 28.sp,
-                            fontWeight: FontWeight.bold,
-                          ),
-                        ),
+                      const AlignedTitleText(
+                        title: "이름을 입력해주세요",
                       ),
                       SizedBox(
                         height: 90.h,
                       ),
-                      SizedBox(
-                        height: 20,
-                        child: Align(
-                          alignment: Alignment.centerLeft,
-                          child: label(),
-                        ),
+                      LabelText(
+                        title: "이름",
+                        focusNode: _focusNode,
+                        controller: nameController,
                       ),
-                      TextField(
+                      SignupTextField(
                         controller: nameController,
                         focusNode: _focusNode,
-                        style: TextStyle(
-                          fontSize: 22.sp,
-                          fontWeight: FontWeight.w500,
-                        ),
-                        decoration: InputDecoration(
-                          contentPadding: const EdgeInsets.only(
-                            bottom: 10,
-                            top: 5,
-                          ),
-                          hintText: _focusNode.hasFocus ? "" : "이름",
-                          hintStyle: TextStyle(
-                            fontSize: 22.sp,
-                            fontWeight: FontWeight.w600,
-                            color: ColorsTheme.gray600,
-                          ),
-                          enabledBorder: const UnderlineInputBorder(
-                            borderSide: BorderSide(
-                              width: 2,
-                              color: ColorsTheme.gray600,
-                            ),
-                          ),
-                          focusedBorder: const UnderlineInputBorder(
-                            borderSide: BorderSide(
-                              width: 2,
-                              color: ColorsTheme.point,
-                            ),
-                          ),
-                        ),
+                        title: "이름",
                       ),
                     ],
                   ),
-                ),
-                Padding(
-                  padding: EdgeInsets.symmetric(
-                    horizontal: 25.w,
-                    vertical: _focusNode.hasFocus ? 20.h : 70.h,
-                  ),
-                  child: ElevatedButton(
-                    onPressed: () => _navigateToNext(nameController),
-                    style: ElevatedButton.styleFrom(
-                      minimumSize: Size(1.sw, 60.h),
-                      backgroundColor: ColorsTheme.point,
-                      shape: RoundedRectangleBorder(
-                        borderRadius: BorderRadius.circular(10),
-                      ),
+                  Padding(
+                    padding: EdgeInsets.symmetric(
+                      vertical: _focusNode.hasFocus ? 20.h : 70.h,
                     ),
-                    child: Text(
-                      "확인",
-                      style: TextStyle(
-                        color: ColorsTheme.white,
-                        fontSize: 18.sp,
-                        fontWeight: FontWeight.bold,
-                      ),
+                    child: SignupButton(
+                      handlePressed: () => _navigateToNext(nameController),
+                      content: "확인",
                     ),
                   ),
-                ),
-              ],
+                ],
+              ),
             ),
           ),
         ),
